No se lograron corregir todos los errores para que compilara

+++ untitled/src/Figura2/Pasaje.java
@@ -15,8 +15,8 @@ public class Pasaje {
         this.viaje = viaje;
     }
 
-    public long getNumero() {
-        return numero;
+    public int getNumero() {
+        return (int)numero;
     }
 
     public int getAsiento() {
